+++ opentelemetry-sdk/CHANGELOG.md
@@ -21,6 +21,10 @@ also modified to suppress telemetry before invoking exporters.
   - There is ability to configure cardinality limits via Instrument 
     advisory. [#2903](https://github.com/open-telemetry/opentelemetry-rust/pull/2903)
 
+- *Breaking* change for custom `MetricReader` authors.
+  The `shutdown_with_timeout` method is added to `MetricReader` trait.
+  `collect` method on `MetricReader` modified to return `OTelSdkResult`.
+
 ## 0.29.0
 
 Released 2025-Mar-21
@@ -78,7 +82,6 @@ Released 2025-Mar-21
   Custom exporters will need to internally synchronize any mutable state, if applicable.
 
 - **Breaking** The `shutdown_with_timeout` method is added to MetricExporter trait. This is breaking change for custom `MetricExporter` authors.
-- **Breaking** The `shutdown_with_timeout` method is added to MetricReader trait. This is breaking change for custom `MetricReader` authors.
 - Bug Fix: `BatchLogProcessor` now correctly calls `shutdown` on the exporter
   when its `shutdown` is invoked.